+++ modules/apps/portal-template/portal-template-test-util/src/main/java/com/liferay/portal/template/test/util/templatemanagerutil/TestTemplate.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.template.bundle.templatemanagerutil;
+package com.liferay.portal.template.test.util.templatemanagerutil;
 
 import com.liferay.portal.kernel.template.Template;
