+++ modules/apps/portal-template/portal-template-test-util/src/main/java/com/liferay/portal/template/test/util/templateresourceloaderutil/TestTemplateResourceLoader.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.template.bundle.templateresourceloaderutil;
+package com.liferay.portal.template.test.util.templateresourceloaderutil;
 
 import com.liferay.portal.kernel.template.TemplateResource;
 import com.liferay.portal.kernel.template.TemplateResourceLoader;
