--- portal-impl/test/integration/com/liferay/portal/kernel/template/bundle/templatehandlerregistryutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for TemplateHandlerRegistryUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
