--- portal-impl/test/integration/com/liferay/portal/kernel/template/bundle/templatemanagerutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for TemplateManagerUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
