--- portal-impl/test/integration/com/liferay/portal/kernel/template/bundle/templateresourceloaderutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for TemplateResourceLoaderUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
